added view tests

+++ gooutsafe/views/users.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, redirect, render_template, request, url_for, flash
-from flask_login import (login_user, login_required)
+from flask_login import (login_user, login_required, current_user)
 
 from gooutsafe.dao.restaurant_manager import RestaurantManager
 from gooutsafe.dao.user_manager import UserManager
@@ -51,12 +51,14 @@ def create_user_type(type_):
 @users.route('/delete_user/<int:id_>', methods=['GET', 'POST'])
 @login_required
 def delete_user(id_):
-    user = UserManager.retrieve_by_id(id_)
-    if user.type == "operator":
-        restaurant = RestaurantManager.retrieve_by_operator_id(id_)
-        RestaurantManager.delete_restaurant(restaurant)
-
-    UserManager.delete_user_by_id(id_)
+    if current_user.id == id_:
+        user = UserManager.retrieve_by_id(id_)
+        if user is not None and user.type == "operator":
+            restaurant = RestaurantManager.retrieve_by_operator_id(id_)
+            RestaurantManager.delete_restaurant(restaurant)
+        
+        UserManager.delete_user_by_id(id_)
+        print("ARRIVATO")
     return redirect(url_for('home.index'))
